Update project to build with submodules
When this project is built by including it in another workspace, instead of building by command line with `carthage`, using the FRAMWORK_SEARCH_PATH of `$(PROJECT_DIR)/Carthage/Build/iOS` can produce an error, since `$(PROJECT_DIR)/Carthage/Build/iOS` might not exist.

To fix this we should use a submodule for our dependencies and include them in the RxKeyboard.xcworkspace. This will allow us to still compile the RxKeyboard library.